+++ app/javascript/src/components/TimeTracking/EntryCard.tsx
