+++ app/javascript/src/components/TimeTracking/WeeklyEntries.tsx
