+++ app/javascript/src/components/TimeTracking/WeeklyEntriesCard.tsx
